+++ src/GafferUI/Tool.cpp
@@ -48,10 +48,14 @@ GAFFER_NODE_DEFINE_TYPE( Tool );
 size_t Tool::g_firstPlugIndex = 0;
 
 Tool::Tool( View *view, const std::string &name )
-	:	Node( name ), m_view( view )
+	:	Node( name ), m_unused( nullptr )
 {
 	storeIndexOfNextChild( g_firstPlugIndex );
 	addChild( new BoolPlug( "active", Plug::In, false ) );
+	view->tools()->addChild( this );
+
+	// Silence Clang warning "private field is not used".
+	(void)m_unused;
 }
 
 Tool::~Tool()
@@ -70,12 +74,40 @@ const Gaffer::BoolPlug *Tool::activePlug() const
 
 View *Tool::view()
 {
-	return m_view;
+	return const_cast<View *>( const_cast<const Tool *>( this )->view() );
 }
 
 const View *Tool::view() const
 {
-	return m_view;
+	auto v = ancestor<View>();
+	if( !v )
+	{
+		throw IECore::Exception( "View not found" );
+	}
+	return v;
+}
+
+bool Tool::acceptsParent( const GraphComponent *potentialParent ) const
+{
+	if( !Node::acceptsParent( potentialParent ) )
+	{
+		return false;
+	}
+
+	if( !potentialParent || potentialParent == parent() )
+	{
+		return true;
+	}
+
+	if( typeId() != staticTypeId() )
+	{
+		// Only accept initial parenting performed in our constructor, before
+		// derived class is initialised (and `typeId()` returns something else).
+		return false;
+	}
+
+	// Only accept parenting to ToolContainers.
+	return IECore::runTimeCast<const ToolContainer>( potentialParent );
 }
 
 //////////////////////////////////////////////////////////////////////////
@@ -130,3 +162,14 @@ void Tool::registeredTools( IECore::TypeId viewType, std::vector<std::string> &t
 		viewType = IECore::RunTimeTyped::baseTypeId( viewType );
 	} while( viewType != (IECore::TypeId)NodeTypeId && viewType != IECore::InvalidTypeId );
 }
+
+//////////////////////////////////////////////////////////////////////////
+// ToolContainer
+//////////////////////////////////////////////////////////////////////////
+
+namespace Gaffer
+{
+
+GAFFER_DECLARECONTAINERSPECIALISATIONS( GafferUI::ToolContainer, ToolContainerTypeId )
+
+} // namespace Gaffer